blurred header on iOS 26
